requirements.txt
langchain
langchain-community
langchain-groq
langchain-huggingface
langgraph
chromadb
sentence-transformers
rank-bm25
ragas
gradio
huggingface-hub
requests
python-dotenv
pandas
tqdm
datasets
